LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_performance_attributes_exist_in_object-expected.txt

Description

This test validates that the window.performance object exists in an object element (but does not validate that their values are correct).



PASS window.performance is defined
PASS window.performance.timing is defined
PASS window.performance.navigation is defined
